Add fixed files
Fix the query for the first question part e. If you compare the amount of forest area lost between 1990 and 2016, to which country's total area in 2016 is it closest to? (The query you wrote doesn't produce right country specifically. It just lists all the countries and their total land areas. We only have to manually inspect the output to know the answer.).
And fix the other query for the third question part  c. If countries were grouped by percent forestation in quartiles, which group had the most countries in it in 2016? (To produce the correct count of countries in the second quartile, you need to prevent World from being counted).